requirements.txt
discord.py
pandas
requests
python-dotenv
web3